requirements/base.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-Bcrypt==0.6.0
Flask-Mail==0.9.1
Flask-Script==2.0.5
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
WTForms==2.0.1
Werkzeug==0.9.6
argparse==1.2.2
blinker==1.3
itsdangerous==0.24
py-bcrypt==0.4
webassets==0.10.1
wsgiref==0.1.2
flask-mongoengine==0.7.1
mongoengine==0.8.7
pymongo==2.7.2
python-instagram==1.3.0
